refactor of the staking markets

+++ defitrack-rest/defitrack-protocol-services/defitrack-quickswap/src/main/java/io/defitrack/protocol/quickswap/staking/QuickswapUserStakingService.kt
@@ -60,7 +60,7 @@ class QuickswapUserStakingService(
                         vaultAddress = pool.address,
                         rate = (quickswapAPRService.getRewardPoolAPR(pool.address) + quickswapAPRService.getLPAPR(
                             stakedToken.address
-                        )).toDouble(),
+                        )),
                         amount = balance
                     )
                 } else {
